update card useguard

+++ src/card/card.service.ts
@@ -3,17 +3,21 @@ import {
   Injectable,
   NotFoundException,
   UnauthorizedException,
+  ForbiddenException,
 } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Card } from "./entities/card.entity";
-import { User } from "../user/entity/user.entity";
 import { Working } from "./entities/working.entity";
+import { User } from "src/user/entity/user.entity";
+import { Board } from "src/board/entities/board.entity";
+import { Colum } from "src/colum/entities/colum.entity";
 import { In, Repository } from "typeorm";
 import {
   CreateCardDto,
   CardDetailsDto,
   UpdateCardDto,
   ReorderCardsDto,
+  MoveCardDto,
 } from "./dto/card.dto";
 
 @Injectable()
@@ -25,10 +29,20 @@ export class CardService {
     private readonly worker_Repository: Repository<Working>,
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    @InjectRepository(Board)
+    private readonly boardRepository: Repository<Board>,
+    @InjectRepository(Colum)
+    private readonly columRepository: Repository<Colum>,
   ) {}
 
-  async createCard(createCardDto: CreateCardDto) {
+  async createCard(user: User, boardId: number, createCardDto: CreateCardDto) {
     const { cardOrder, title, desc, color } = createCardDto;
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
 
     const newCard = this.cardRepository.create({
       cardOrder: cardOrder,
@@ -36,7 +50,6 @@ export class CardService {
       desc: desc,
       color: color,
     });
-
     return this.cardRepository.save(newCard);
   }
 
@@ -70,7 +83,18 @@ export class CardService {
     return cardDetails;
   }
 
-  async createCardWorker_(cardId: number, UserId: number): Promise<void> {
+  async createCardWorker_(
+    user: User,
+    boardId: number,
+    cardId: number,
+    UserId: number,
+  ): Promise<void> {
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
     const card = await this.cardRepository.findOne({
       where: {
         id: cardId,
@@ -84,7 +108,7 @@ export class CardService {
     });
 
     if (!card || !user) {
-      throw new NotFoundException("User or Card not found");
+      throw new NotFoundException("유저 또는 카드를 찾을 수 없습니다.");
     }
 
     const worker_ = this.worker_Repository.create({
@@ -95,14 +119,36 @@ export class CardService {
     await this.worker_Repository.save(worker_);
   }
 
-  async removeCardWorker_(cardId: number, userId: number): Promise<void> {
+  async removeCardWorker_(
+    user: User,
+    boardId: number,
+    cardId: number,
+    workerId: number,
+  ): Promise<void> {
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
     await this.worker_Repository.delete({
       card: { id: cardId },
-      user: { id: userId },
+      user: { id: workerId },
     });
   }
 
-  async updateCard(updateCardDto: UpdateCardDto, cardId: number) {
+  async updateCard(
+    user: User,
+    boardId: number,
+    updateCardDto: UpdateCardDto,
+    cardId: number,
+  ) {
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
     const { title, desc, color } = updateCardDto;
     const card = await this.cardRepository.findOne({
       where: {
@@ -111,7 +157,7 @@ export class CardService {
     });
 
     if (!card) {
-      throw new NotFoundException("Card not found");
+      throw new NotFoundException("카드를 찾을 수 없습니다.");
     }
 
     card.title = title ? title : card.title;
@@ -121,11 +167,27 @@ export class CardService {
     return this.cardRepository.save(card);
   }
 
-  async deleteCard(cardId: number): Promise<void> {
+  async deleteCard(user: User, boardId: number, cardId: number): Promise<void> {
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
     await this.cardRepository.delete(cardId);
   }
 
-  async reorderCards(reorderCardsDto: ReorderCardsDto): Promise<void> {
+  async reorderCards(
+    user: User,
+    boardId: number,
+    reorderCardsDto: ReorderCardsDto,
+  ): Promise<void> {
+    const userId = user.id;
+    const isMember = await this.isMemberOfBoard(userId, boardId);
+
+    if (!isMember) {
+      throw new ForbiddenException("해당 보드의 멤버가 아닙니다.");
+    }
     const { cardIds } = reorderCardsDto;
     const cards = await this.cardRepository.find({
       where: {
@@ -145,5 +207,52 @@ export class CardService {
         return this.cardRepository.save(card);
       }),
     );
+  }
+
+  async moveCard(moveCardDto: MoveCardDto): Promise<void> {
+    const { cardId, newColumId } = moveCardDto;
+
+    const card = await this.cardRepository.findOne({
+      where: {
+        id: cardId,
+      },
+    });
+
+    if (!card) {
+      throw new NotFoundException("Card not found");
+    }
+
+    const newColum = await this.columRepository.findOne({
+      where: {
+        id: newColumId,
+      },
+    });
+
+    if (!newColum) {
+      throw new NotFoundException("Target list not found");
+    }
+
+    card.colum = newColum;
+
+    this.cardRepository.save(card);
+  }
+
+  ////////////////////////////////////privatefunction///////////////////////////////////////
+  private async isMemberOfBoard(
+    userId: number,
+    boardId: number,
+  ): Promise<boolean> {
+    const board = await this.boardRepository
+      .createQueryBuilder("board")
+      .innerJoinAndSelect("board.members", "members")
+      .innerJoinAndSelect("members.user", "user")
+      .where("board.id = :boardId", { boardId })
+      .getOne();
+
+    if (!board) {
+      return false;
+    }
+
+    return board.members.some((member) => member.user.id === userId);
   }
 }
